requirements.txt
albumentations==2.0.8
datasets==4.0.0
evaluate==0.4.5
huggingface-hub==0.34.4
isort==6.0.1
joblib==1.5.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
numpy==2.1.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pathlib2==2.3.7.post1
pfzy==0.3.4
pillow==11.0.0
pylint==3.3.8
PyMuPDF==1.26.3
pyparsing==3.2.3
pytesseract==0.3.13
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy
streamlit==1.48.1
streamlit-image-zoom==0.0.4
torch==2.8.0
torchvision==0.23.0
transformers==4.55.2
ultralytics==8.3.184
streamlit_image_coordinates==0.4.0